George gave idea for better refactor
